Hopefully fixed hanging bug

+++ bootloader/src/bootloader.c
@@ -256,8 +256,8 @@ void handle_update(void)
     uart_writeb(HOST_UART, FRAME_OK);
 
     // Now we get and decrypt the firmware and check that its signed
-    uart_read(HOST_UART, firmbuff, size);
-    
+    uart_read(HOST_UART, firmbuff, (uint32_t)size);
+
     // Acknowledge
     uart_writeb(HOST_UART, FRAME_OK);
